refactor(qbittorrent-e2e): extract workspace module
