+++ src/console/ci/qbittorrent/mod.rs
@@ -1,3 +1,4 @@
 pub mod bencode;
 pub mod qbittorrent_client;
 pub mod runner;
+pub mod workspace;
